# Optimization: Avoid redundant work when checking completed works