--- a/src/lib/mina_intf/dune
+++ b/src/lib/mina_intf/dune
@@ -26,6 +26,7 @@
    verifier
    rose_tree
    mina_ledger
+   transaction_snark_work
    )
  (instrumentation (backend bisect_ppx))
  (preprocess (pps ppx_jane ppx_version ppx_deriving.std ppx_deriving_yojson)))