--- a/src/lib/mina_lib/tests/tests.ml
+++ b/src/lib/mina_lib/tests/tests.ml
@@ -279,8 +279,9 @@ let%test_module "Epoch ledger sync tests" =
             (make_dirname "persistent_frontier_location")
           ~frontier_broadcast_pipe:
             (frontier_broadcast_pipe_r, frontier_broadcast_pipe_w)
-          ~catchup_mode:`Normal ~network_transition_reader:block_reader
-          ~producer_transition_reader ~most_recent_valid_block ~notify_online ()
+          ~get_completed_work:(Fn.const None) ~catchup_mode:`Normal
+          ~network_transition_reader:block_reader ~producer_transition_reader
+          ~most_recent_valid_block ~notify_online ()
       in
       let%bind () = Ivar.read initialization_finish_signal in
       let network_peer =