The patch submitted here was originally made and tested against the `release/2.0.0` branch (connected to berkeleynet and in a private cluster) when investigating where the node spends time when processing a block.

Submitted as `Draft` for now because there are probably some tests need to be added for this change (in particular, making sure that only completed work that exactly matches the one in the block is considered as valid for skipping the verification).

The core logic is in the changes to the `src/lib/staged_ledger/staged_ledger.ml` file. The rest of the files had to be changed to make the `get_completed_work` function available to the block processing code that verifies the work included in the block. If there is a better way that doesn't require passing this around please let me know.

### Explain your changes:

When applying a block, when works are included, the biggest contributor to the time it takes to process the block is the "check completed works" step (on which the included completed works are verified).

But most of the time the completed work that needs to be verified has already been verified in the SNARK pool. This optimization takes advantage of that by comparing and filtering out all the completed works included in a block that are already present in the SNARK pool. As a result, the amount of work that needs to be verified is reduced (quite often to 0).

### Explain how you tested your changes:

We are running two servers extended with internal tracing [here](https://openmina-tracing.web.app/tracing/blocks), one unoptimized, and the other optimized (with this patch and #12521).

For block 7881, which includes 76 completed works, here are the times for the unpatched and patched nodes:

Unpatched (8.86s) / Block Total (12.03s) |  Patched (0.015s) / Block Total (1.5s)
:-------------------------:|:-------------------------:
![check-completed-works-7881-unpatched](https://user-images.githubusercontent.com/13744/213319169-3ef81327-75b5-480f-991a-49e74ffd43f2.png) | ![check-completed-works-7881-patched](https://user-images.githubusercontent.com/13744/213319226-56b51eb9-ca6d-4ba5-ba68-72ceb2685f1a.png)

For block 8024, which includes 126 completed works:

Unpatched (14.32s) / Block Total (16.49s) |  Patched (0.025s) / Block Total (1.48s)
:-------------------------:|:-------------------------:
![check-completed-works-8024-unpatched](https://user-images.githubusercontent.com/13744/213444833-7c397c46-2775-450f-9973-0b0bed809ca5.png) | ![check-completed-works-8024-patched](https://user-images.githubusercontent.com/13744/213444860-73fdcb10-79ce-4ffd-bd4e-3691d17a0ec1.png)

In these cases, and because on berkeleynet most of the works are completed by the same prover, the checks were completely skipped.

But with multiple provers, the chances of the node having in its SNARK pool a completed work matching the statement, but with a different prover or fee increases (and in such cases the work verification will not be skipped), because the node will only keep the completed work with the lowest fee (and if two works with the same fee reach the node, the first one will be kept).

Checklist:

- [ ] Modified the current draft of release notes with details on what is completed or incomplete within this project
- [ ] Document code purpose, how to use it
  - Mention expected invariants, implicit constraints
- [ ] Tests were added for the new behavior
  - Document test purpose, significance of failures
  - Test names should reflect their purpose
- [ ] All tests pass (CI will check this if you didn't)
- [ ] Serialized types are in stable-versioned modules
- [ ] Does this close issues? List them

* Closes #0000