@tizoc 2023-02-01
There is a more general (and maybe even cleaner) variant that would also take advantage of works seen when processing other blocks. Instead of directly querying the SNARK pool, a new cache could be added to keep track of all the proofs that are known to be valid. The block validator would then query (and update) this cache instead.

Another alternative is for this cache to be maintained by the verifier process. The disadvantage of doing that is that:
- it doesn't work as well when multiple verifiers are added (unless synchronization between verifiers is implemented too)
- if the verifier process is busy verifying other proofs, the caller needs to wait
- probably ends up duplicating more data between processes