@mrmr1993 2023-09-06
Looks like you have a failure in the tests (`src/lib/mina_lib/tests/tests.ml`).